requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-import-export==2.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
MarkupPy==1.14
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0
